+++ patterns--result/Patterns.Result/ResultExtensions.cs
@@ -1,8 +1,0 @@
-namespace Patterns.Result;
-
-public static class ResultExtensions
-{
-    public static bool IsFailure(this Result result) => !result.IsSuccess;
-
-    public static int StatusCode(this Result result) => result.Status.ToStatusCode();
-}